refactor: sanitizer, getfonts 함수 분리, 모달 수정
